% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_allbus.R
\name{load_allbus}
\alias{load_allbus}
\title{Load ALLBUS data}
\usage{
load_allbus(path)
}
\arguments{
\item{path}{Pfad zur ALLBUS-Datei (.sav, .dta)}
}
\value{
Dataframe
}
\description{
Lädt den ALLBUS-Datensatz von einem lokalen Pfad.
}
