% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:plot_eastwest.R, R/plot_eastwest.R
\name{plot_eastwest}
\alias{plot_eastwest}
\title{Plot East/West Verteilung}
\usage{
plot_eastwest(data, var = "eastwest")

plot_eastwest(data, var = "eastwest")
}
\arguments{
\item{data}{Ein Dataframe (ALLBUS Datensatz)}

\item{var}{Name der East/West Spalte (z. B. "eastwest")}
}
\description{
Plottet die Häufigkeit der East/West Variable (1 = East, 2 = West).

Plot East/West Verteilung
}
\details{
Plottet die Häufigkeit der East/West Variable (1 = Ost, 2 = West).
}
