% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_allbus.R
\name{summary_allbus}
\alias{summary_allbus}
\title{Fasse den ALLBUS Datensatz zusammen}
\usage{
summary_allbus(data)
}
\arguments{
\item{data}{Ein Dataframe (ALLBUS Datensatz)}
}
\value{
Ein Tibble mit den Zusammenfassungen
}
\description{
Zeigt die Anzahl Zeilen, Spalten und einfache Statistiken.
}
